Move generic functions to util
